requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.1
idna==2.8
idna-ssl==1.1.0
multidict==4.7.4
typing-extensions==3.7.4.1
websockets==8.1
yarl==1.4.2
